Add glade file, remove inline UI code
This commit will remove the inline UI code in main.py and replace
it with a Glade XML file aptly named owlywindow.glade. This markup
has a tree representing all of the objects in the window(s) that we
then can load and manipulate with out code in main.py - handy huh!