Fix: OwlyWindow.quit() not being called on window destroy event (#15)
This commit fixes changes the event-handler that calls OwlyWindow.quit()
from `destroy-event` to `destroy`. Also change quit() to accept a list
of args and kwargs since we don't actually need to use them right now.
Probably not the best way but we don't care so as long as quit gets
called.